requirements.txt
dnspython
django
gunicorn
django-heroku
django-crispy-forms
dj-database-url
whitenoise
mongoengine
django_mongoengine
djangorestframework
django-rest-framework-mongoengine
django-filter
markdown